Refactor custom property handling and registration
